add enemy boss class

+++ src/secondGame/FastEnemy.java
@@ -18,18 +18,18 @@ public class FastEnemy extends GameObject {
         y+=velY;
         if(x<=0||x>=Game.WIDTH-32) velX*=-1;
         if(y<=0||y>=Game.HEIGHT-32) velY*=-1;
-        handler.addObject(new Trail(x,y,ID.Trail,Color.CYAN,16,16,0.01f,handler));
+        handler.addObject(new Trail((int)x,(int)y,ID.Trail,Color.CYAN,16,16,0.01f,handler));
     }
 
     @Override
     public void render(Graphics g) {
         g.setColor(Color.CYAN);
-        g.fillRect(x,y,16,16);
+        g.fillRect((int)x,(int)y,16,16);
 
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x,y,16,16);
+        return new Rectangle((int)x,(int)y,16,16);
     }
 }
