+++ src/secondGame/HUD.java
@@ -17,8 +17,8 @@ public class HUD {
     public void render(Graphics g){
         g.setColor(Color.gray);
         g.fillRect(15,15,200,32);
-        g.setColor(new Color(75,greenValue,0));
-        g.fillRect(15,15,HEALTH*2,32);
+        g.setColor(new Color(75,(int)greenValue,0));
+        g.fillRect(15,15,(int)HEALTH*2,32);
         g.setColor(Color.white);
         g.drawRect(15,15,200,32);
